requirements.txt
Flask==2.3.3
keras==2.7.0
numpy==1.24.4
opencv_python==4.8.0.76
Pillow==10.0.1
Pillow==10.0.1
Requests==2.31.0
tensorflow_gpu==2.7.0
torch==1.12.0
ultralytics==8.0.180
Werkzeug==2.3.7
